slicetype: upgrade frame cost variables to uint64_t

+++ source/common/lowres.cpp
@@ -127,7 +127,7 @@ void Lowres::init(TComPicYuv *orig, int poc, int type, int bframes)
     sliceType = type;
     frameNum = poc;
     leadingBframes = 0;
-    satdCost = -1;
+    satdCost = (uint64_t)-1;
     memset(costEst, -1, sizeof(costEst));
 
     if (qpAqOffset && invQscaleFactor)
